# cmd/api/docs/swagger.yaml
definitions:
...
  models.OTPData:
    properties:
      phone:
        type: string
    required:
    - phone
    type: object
  models.Order:
    properties:
      address_id:
        type: integer
      coupon_id:
        type: integer
      payment_id:
        type: integer
      user_id:
        type: integer
    type: object
...
  models.VerifyData:
    properties:
      code:
        type: string
      phone:
        type: string
    required:
    - code
    - phone
    type: object
  response.Response:
    properties:
      data: {}
      error: {}
      message:
        type: string
      status_code:
        type: integer
    type: object
info:
  contact:
    name: API Support
  description: Watch Hive is an E-commerce platform to purchase Watch
  title: Go + Gin E-Commerce API Watch Hive
  version: 1.0.0
paths:
...
  /admin/orders:
    get:
      consumes:
      - application/json
      description: Retrieve all orders for admin view.
      parameters:
      - description: Page number for pagination
        in: query
        name: page
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Successfully retrieved all admin orders
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Invalid input or error retrieving orders
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerTokenAuth: []
      summary: Get Admin Orders
      tags:
      - Admin Order Management
  /admin/orders/approve:
    post:
      consumes:
      - application/json
      description: Approve an order by updating its status.
      parameters:
      - description: Order ID to be approved
        in: query
        name: order_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Successfully approved the order
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Invalid input or error approving the order
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerTokenAuth: []
      summary: Approve Order
      tags:
      - Admin Order Management
  /admin/orders/print-invoice:
    get:
      consumes:
      - application/json
      description: Generate and download the invoice for a specific order.
      parameters:
      - description: Order ID for which the invoice should be generated
        in: query
        name: order_id
        required: true
        type: string
      produces:
      - application/pdf
      responses:
        "200":
          description: Invoice PDF
          schema:
            type: file
        "400":
          description: Error in processing the request
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Error generating or downloading the invoice
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerTokenAuth: []
      summary: Print Invoice
      tags:
      - User Order Management
  /admin/orders/return:
    post:
      consumes:
      - application/json
      description: Initiate the return process for an order by updating its status.
      parameters:
      - description: Order ID to be returned
        in: query
        name: order_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Successfully initiated the return process
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Error initiating the return process
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerTokenAuth: []
      summary: Return Order
      tags:
      - User Order Management
...
  /orders:
    get:
      consumes:
      - application/json
      description: Retrieve orders based on the provided order ID.
      parameters:
      - description: Order ID
        in: query
        name: order_id
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Orders retrieved successfully
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Invalid input or error retrieving orders
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerTokenAuth: []
      summary: Get Orders
      tags:
      - User Order Management
  /orders/all:
    get:
      consumes:
      - application/json
      description: Retrieve all orders for the authenticated user.
      parameters:
      - description: Page number for pagination (default is 1)
        in: query
        name: page
        type: integer
      - description: Number of items per page (default is 10)
        in: query
        name: count
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Successfully retrieved all orders
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Invalid input or error retrieving orders
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerTokenAuth: []
      summary: Get All Orders
      tags:
      - User Order Management
  /orders/cancel:
    post:
      consumes:
      - application/json
      description: Cancel an order based on the provided order ID.
      parameters:
      - description: Order ID
        in: query
        name: order_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Order successfully canceled
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Invalid input or error canceling the order
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerTokenAuth: []
      summary: Cancel Order
      tags:
      - User Order Management
  /orders/order-items-from-cart:
    post:
      consumes:
      - application/json
      description: Place an order for items in the user's cart.
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      - description: Order details including address, payment method, and coupon ID
        in: body
        name: order
        required: true
        schema:
          $ref: '#/definitions/models.Order'
      produces:
      - application/json
      responses:
        "200":
          description: Order placed successfully
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Invalid input or error placing the order
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerTokenAuth: []
      summary: Order Items From Cart
      tags:
      - User Order Management
  /otp/send:
    post:
      consumes:
      - application/json
      description: Send OTP to the specified phone number for verification.
      parameters:
      - description: Phone number details in JSON format
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.OTPData'
      produces:
      - application/json
      responses:
        "200":
          description: OTP sent successfully
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Fields provided in wrong format or Could not send OTP
          schema:
            $ref: '#/definitions/response.Response'
      summary: Send OTP
      tags:
      - User
  /otp/verify:
    post:
      consumes:
      - application/json
      description: Verify the provided OTP code for user authentication.
      parameters:
      - description: Verification code details in JSON format
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.VerifyData'
      produces:
      - application/json
      responses:
        "200":
          description: Successfully verified OTP
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Fields provided in wrong format or Could not verify OTP
          schema:
            $ref: '#/definitions/response.Response'
      summary: Verify OTP
      tags:
      - User
...
securityDefinitions:
  BearerTokenAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"